backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
groq
sentence-transformers
pgvector
sqlalchemy
psycopg2-binary>=2.9.11

redis==5.0.1
httpx
python-dotenv
pydantic-settings
